Deleted old tests and renamed new_tests
also changed paths in scripts
