--- new_tests/evaluate.toml
@@ -1,40 +1,0 @@
-# This is a TOML document.
-
-title = "Evaluation configuration"
-
-[mode]
-quiet = false
-gpu = false
-
-[paths]
-data_path = "./new_tests/build/data/"
-model_path = "./new_tests/model/model_eval.model"
-model_path_2 = "none"
-norm_path = "none"
-
-[general]
-fourier = true
-amp_phase = true
-source_list = false
-arch_name = "SRResNet_bigger"
-arch_name_2 = "none"
-output_format = "png"
-diff = true
-
-[inspection]
-visualize_prediction = true
-visualize_source_reconstruction = true
-visualize_contour = false
-visualize_dynamic_range = false
-visualize_blobs = false
-visualize_ms_ssim = false
-random = true
-num_images = 10
-
-[eval]
-batch_size = 100
-evaluate_viewing_angle = false
-evaluate_dynamic_range = false
-evaluate_ms_ssim = false
-evaluate_mean_diff = false
-evaluate_area = false
