+++ tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 import shutil
 
 
-@pytest.yield_fixture(autouse=True, scope='session')
+@pytest.fixture(autouse=True, scope='session')
 def test_suite_cleanup_thing():
     yield
