+++ tests/simulate.toml
@@ -1,0 +1,47 @@
+# This is a TOML document.
+
+title = "Simulation configuration"
+
+[mode]
+quiet = true
+
+[paths]
+data_path = "./tests/build/data/"
+data_format = "h5"
+
+[mnist]
+simulate = false
+resource = "../../resources/mnist.pkl.gz"
+
+[gaussians]
+simulate = true
+pointsources = false
+num_pointsources = 5
+pointlike_gaussians = false
+num_pointlike_gaussians = 5
+extended_gaussians = true
+num_components = [4, 10]
+
+
+[image_options]
+bundles_train = 1
+bundles_valid = 1
+bundles_test = 1
+bundle_size = 100
+img_size = 63
+noise = true
+noise_level = 5
+
+[sampling_options]
+fourier = true
+real_imag = false
+amp_phase = true
+specific_mask = true
+lon = -80
+lat = 50
+steps = 50
+antenna_config = "vlba"
+compressed = false
+keep_fft_files = true
+source_list = false
+interpolation = true
